cli/pyproject.toml
[project]
name = "plex-wrapped"
version = "0.1.0"
description = "Self-hostable Spotify Wrapped for Plex servers"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "typer>=0.9.0",
    "rich>=13.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "plexapi>=4.15.0",
    "anthropic>=0.18.0",
    "openai>=1.12.0",
    "pyyaml>=6.0.0",
    "textual>=0.47.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
plex-wrapped = "plex_wrapped.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/plex_wrapped"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
asyncio_mode = "auto"

[dependency-groups]
dev = [
    "pytest-asyncio>=1.3.0",
]
